algo/tests: move tests around
